Move and refactor `Scheme` and rename parser file
